move down diagnostic analyzer executor down to workspace layer since all requires one now lives in workspace layer.
